+++ R/estimators.R
@@ -36,8 +36,8 @@
 #'   criteria).
 #'
 #' @references
-#' Arora, P. and Wagle, R. (2026). Partial Homogeneity in Staggered
-#' Difference-in-Differences. Eq. (12), (50), (55) and Appendix B.
+#' Arora, P. and Wagle, R. (2026). Eq. (12), (50), (55) and Appendix B.
+#' See \code{citation("phdid")} for the full reference.
 #'
 #' @export
 #' @examples